move to algoOp
