+++ packages/algob/src/lib/algo-operator.ts
@@ -32,7 +32,6 @@ export function createAlgoOperator (network: Network): AlgoOperator {
 
 export interface AlgoOperator {
   algodClient: algosdk.Algodv2
-  compileOp: CompileOp
   deployASA: (
     name: string, asaDef: ASADef, flags: ASADeploymentFlags, accounts: Accounts, txWriter: txWriter
   ) => Promise<ASAInfo>
@@ -45,6 +44,7 @@ export interface AlgoOperator {
   optInToASAMultiple: (
     asaName: string, asaDef: ASADef, flags: ASADeploymentFlags, accounts: Accounts, assetIndex: number
   ) => Promise<void>
+  getLogicSignature: (name: string, scParams: Object) => Promise<Object | undefined>
 }
 
 export class AlgoOperatorImpl implements AlgoOperator {
@@ -233,6 +233,16 @@ export class AlgoOperatorImpl implements AlgoOperator {
       logicSignature: lsig,
       confirmedRound: confirmedTxn[confirmedRound]
     };
+  }
+
+  async getLogicSignature(name: string, scParams: Object): Promise<Object | undefined> {
+    const result = await this.compileOp.readArtifact(name);
+    if(result === undefined)
+      return undefined;
+    const programb64 = result.compiled;
+    const program = new Uint8Array(Buffer.from(programb64, "base64"));
+    const lsig = algosdk.makeLogicSig(program, scParams);
+    return lsig;
   }
 
   private async ensureCompiled (name: string, force: boolean): Promise<ASCCache> {
